+++ src/libs/AxelarAssetTransfer.ts
@@ -12,17 +12,32 @@ import { CLIENT_API_GET_OTC, CLIENT_API_POST_TRANSFER_ASSET, OTC } from "../serv
 import { RestService, SocketService } from "../services";
 import {
   createWallet,
-  validateChainIdentifier,
+  throwIfInvalidChainIds,
   validateDestinationAddressByChainName,
 } from "../utils";
-import { getConfigs } from "../constants";
-import { AxelarAssetTransferConfig, Environment } from "./types";
+import { EnvironmentConfigs, getConfigs } from "../constants";
+import { AxelarAssetTransferConfig, Environment, GasToken, isNativeToken } from "./types";
 import DepositReceiver from "../../artifacts/contracts/deposit-service/DepositReceiver.sol/DepositReceiver.json";
 import ReceiverImplementation from "../../artifacts/contracts/deposit-service/ReceiverImplementation.sol/ReceiverImplementation.json";
 import s3 from "./TransactionRecoveryApi/constants/s3";
 
 import { constants } from "ethers";
+import { ChainInfo } from "../chains/types";
+import { loadChains } from "../chains";
+import { AxelarQueryAPI } from "./AxelarQueryAPI";
 const { HashZero } = constants;
+
+interface GetDepositAddressParams {
+  fromChain: string;
+  toChain: string;
+  destinationAddress: string;
+  asset: string;
+  options?: {
+    _traceId?: string;
+    shouldUnwrapIntoNative: boolean;
+    refundAddress: string;
+  };
+}
 
 export class AxelarAssetTransfer {
   readonly environment: Environment;
@@ -30,6 +45,7 @@ export class AxelarAssetTransfer {
 
   readonly api: RestService;
   readonly depositServiceApi: RestService;
+  readonly axelarQueryApi: AxelarQueryAPI;
   private gasReceiverContract: Record<string, string> = {};
   private depositServiceContract: Record<string, string> = {};
   private evmDenomMap: Record<string, string> = {};
@@ -46,6 +62,13 @@ export class AxelarAssetTransfer {
 
     this.api = new RestService(this.resourceUrl);
     this.depositServiceApi = new RestService(configs.depositServiceUrl);
+
+    const links: EnvironmentConfigs = getConfigs(config.environment);
+    this.axelarQueryApi = new AxelarQueryAPI({
+      environment: config.environment,
+      axelarRpcUrl: links.axelarRpcUrl,
+      axelarLcdUrl: links.axelarLcdUrl,
+    });
   }
 
   async getDepositAddressForNativeWrap(
@@ -54,7 +77,16 @@ export class AxelarAssetTransfer {
     destinationAddress: string,
     refundAddress?: string
   ): Promise<string> {
-    refundAddress = refundAddress || (await this.getGasReceiverContractAddress(fromChain));
+    await throwIfInvalidChainIds([fromChain, toChain], this.environment);
+    await this.axelarQueryApi.throwIfInactiveChains([fromChain, toChain]);
+
+    refundAddress =
+      refundAddress ||
+      (await this.axelarQueryApi.getContractAddressFromConfig(
+        fromChain,
+        "default_refund_collector"
+      ));
+
     const { address } = await this.getDepositAddressFromRemote(
       "wrap",
       fromChain,
@@ -82,9 +114,18 @@ export class AxelarAssetTransfer {
     fromChain: string,
     toChain: string,
     destinationAddress: string,
+    fromChainModule: "evm" | "axelarnet" = "evm",
     refundAddress?: string
   ): Promise<string> {
-    refundAddress = refundAddress || (await this.getGasReceiverContractAddress(fromChain));
+    await throwIfInvalidChainIds([fromChain, toChain], this.environment);
+    await this.axelarQueryApi.throwIfInactiveChains([fromChain, toChain]);
+
+    refundAddress =
+      refundAddress ||
+      (await this.axelarQueryApi.getContractAddressFromConfig(
+        fromChainModule === "evm" ? fromChain : toChain,
+        "default_refund_collector"
+      ));
 
     const { address: unwrapAddress } = await this.getDepositAddressFromRemote(
       "unwrap",
@@ -126,7 +167,18 @@ export class AxelarAssetTransfer {
     hexSalt: string
   ): Promise<{ address: string }> {
     const endpoint = wrapOrUnWrap === "wrap" ? "/deposit/wrap" : "/deposit/unwrap";
-    return await this.depositServiceApi
+
+    if (fromChain) {
+      await throwIfInvalidChainIds([fromChain], this.environment);
+      await this.axelarQueryApi.throwIfInactiveChains([fromChain]);
+    }
+
+    if (toChain) {
+      await throwIfInvalidChainIds([toChain], this.environment);
+      await this.axelarQueryApi.throwIfInactiveChains([toChain]);
+    }
+
+    return this.depositServiceApi
       .post(endpoint, {
         salt: hexSalt,
         source_chain: fromChain,
@@ -146,6 +198,9 @@ export class AxelarAssetTransfer {
     refundAddress: string,
     hexSalt: string
   ) {
+    await throwIfInvalidChainIds([fromChain, toChain], this.environment);
+    await this.axelarQueryApi.throwIfInactiveChains([fromChain, toChain]);
+
     const receiverInterface = new Interface(ReceiverImplementation.abi);
     const functionData =
       wrapOrUnWrap === "wrap"
@@ -160,7 +215,10 @@ export class AxelarAssetTransfer {
           ]);
 
     const address = getCreate2Address(
-      await this.getDepositServiceContractAddress(fromChain),
+      await this.axelarQueryApi.getContractAddressFromConfig(
+        wrapOrUnWrap === "wrap" ? fromChain : toChain,
+        "deposit_service"
+      ),
       hexSalt,
       keccak256(
         solidityPack(
@@ -176,28 +234,41 @@ export class AxelarAssetTransfer {
   }
 
   /**
-   *
-   * @param fromChain Source chain identifier eg: avalanche, moonbeam ethereum-2, terra-2 ...
-   * @param toChain Destination chain identifier eg: avalanche, moonbeam ethereum-2, terra-2 ...
-   * @param destinationAddress Address where the asset should be transferred to on the destination chain
-   * @param asset Asset denomination eg: uausdc, uaxl ...
-   * @param options
-   * @returns
+   * @param {Object}  requestParams
+   * @param {string}  requestParams.fromChain - Source chain identifier eg: avalanche, moonbeam ethereum-2, terra-2 ...
+   * @param {string}  requestParams.toChain - Destination chain identifier eg: avalanche, moonbeam ethereum-2, terra-2 ...
+   * @param {string}  requestParams.destinationAddress - Address where the asset should be transferred to on the destination chain
+   * @param {string}  requestParams.asset - Asset denomination eg: uausdc, uaxl ... If the asset specific is native cxy (e.g. ETH, AVAX, etc), the ERC20 version of the asset will appear on the dest chain
+   * @param {Object}  requestParams.options
+   * @param {string}  requestParams.options._traceId
+   * @param {boolean} requestParams.options.shouldUnwrapIntoNative - when sending wrapped native asset back to its home chain (e.g. WETH back to Ethereum), specify "true" to receive native ETH; otherwise will received ERC20 version
+   * @param {string}  requestParams.options.refundAddress - recipient where funds can be refunded if wrong ERC20 asset is deposited; ONLY AVAILABLE FOR WRAP/UNWRAP SERVICE
    */
   public async getDepositAddress(
-    fromChain: string,
-    toChain: string,
-    destinationAddress: string,
-    asset: string,
-    options?: {
-      _traceId: string;
-    }
+    requestParamsOrFromChain: GetDepositAddressParams | string,
+    _toChain?: string,
+    _destinationAddress?: string,
+    _asset?: string,
+    _options?: any
   ): Promise<string> {
+    let fromChain: string, toChain: string, destinationAddress: string, asset: string, options: any;
+
+    if (typeof requestParamsOrFromChain === "string") {
+      fromChain = requestParamsOrFromChain;
+      toChain = _toChain as string;
+      destinationAddress = _destinationAddress as string;
+      asset = _asset as string;
+      options = _options;
+    } else {
+      ({ fromChain, toChain, destinationAddress, asset, options } =
+        requestParamsOrFromChain as GetDepositAddressParams);
+    }
+
     // use trace ID sent in by invoking user, or otherwise generate a new one
     const traceId = options?._traceId || uuidv4();
 
     // validate chain identifiers
-    await this.validateChainIdentifiers(fromChain, toChain);
+    await throwIfInvalidChainIds([fromChain, toChain], this.environment);
 
     // verify destination address format
     const isDestinationAddressValid = await validateDestinationAddressByChainName(
@@ -207,6 +278,37 @@ export class AxelarAssetTransfer {
     );
     if (!isDestinationAddressValid)
       throw new Error(`Invalid destination address for chain ${toChain}`);
+
+    const chainList: ChainInfo[] = await loadChains({ environment: this.environment });
+
+    const srcChainInfo = chainList.find(
+      (chainInfo) => chainInfo.id === fromChain.toLowerCase()
+    ) as ChainInfo;
+    if (!srcChainInfo) throw new Error("cannot find chain" + fromChain);
+    const destChainInfo = chainList.find(
+      (chainInfo) => chainInfo.id === toChain.toLowerCase()
+    ) as ChainInfo;
+    if (!destChainInfo) throw new Error("cannot find chain" + toChain);
+
+    /**if user has selected native cxy, e.g. ETH, AVAX, etc, assume it is to be wrapped into ERC20 on dest chain */
+    if (isNativeToken(srcChainInfo.chainName.toLowerCase(), asset as GasToken)) {
+      return this.getDepositAddressForNativeWrap(
+        fromChain,
+        toChain,
+        destinationAddress,
+        options?.refundAddress
+      );
+    }
+    /**if user has selected native cxy wrapped asset, e.g. WETH, WAVAX, and selected to unwrap it */
+    if (destChainInfo.nativeAsset.includes(asset as string) && options?.shouldUnwrapIntoNative) {
+      return this.getDepositAddressForNativeUnwrap(
+        fromChain,
+        toChain,
+        destinationAddress,
+        srcChainInfo.module,
+        options.refundAddress
+      );
+    }
 
     // auth/rate limiting
     const wallet = createWallet();
@@ -255,6 +357,9 @@ export class AxelarAssetTransfer {
   ): Promise<string> {
     type RoomIdResponse = Record<"data", Record<"roomId", string>>;
 
+    await throwIfInvalidChainIds([fromChain, toChain], this.environment);
+    await this.axelarQueryApi.throwIfInactiveChains([fromChain, toChain]);
+
     const payload = {
       fromChain,
       toChain,
@@ -281,6 +386,8 @@ export class AxelarAssetTransfer {
     destinationChain: string,
     destinationAddress: string
   ): Promise<string> {
+    await throwIfInvalidChainIds([sourceChain, destinationChain], this.environment);
+    await this.axelarQueryApi.throwIfInactiveChains([sourceChain, destinationChain]);
     const { newRoomId } = await this.getSocketService()
       .joinRoomAndWaitForEvent(roomId, sourceChain, destinationChain, destinationAddress)
       .catch((error) => {
@@ -298,63 +405,32 @@ export class AxelarAssetTransfer {
     return JSON.parse(roomId)?.depositAddress;
   }
 
-  async getGasReceiverContractAddress(chainName: string): Promise<string> {
-    if (!this.gasReceiverContract[chainName]) {
-      this.gasReceiverContract[chainName] = await this.getStaticInfo()
-        .then((body) => {
-          return body.assets.network[chainName.toLowerCase()]?.gas_service;
-        })
-        .catch((e) => undefined);
-    }
-    return this.gasReceiverContract[chainName];
-  }
+  async getERC20Denom(chainId: string): Promise<string> {
+    const chainList: ChainInfo[] = await loadChains({ environment: this.environment });
+    const chainName = chainList.find(
+      (chainInfo) => chainInfo.id === chainId?.toLowerCase()
+    )?.chainName;
+    if (!chainName) throw new Error(`Chain id ${chainId} does not fit any supported chain`);
 
-  async getERC20Denom(chainName: string): Promise<string> {
     if (!this.evmDenomMap[chainName.toLowerCase()]) {
       const staticInfo = await this.getStaticInfo();
       const denom = staticInfo.chains[chainName.toLowerCase()]?.nativeAsset[0];
       if (denom) {
         this.evmDenomMap[chainName.toLowerCase()] = denom;
+      } else {
+        throw new Error(`Asset denom for ${chainId} not found`);
       }
       return denom;
     }
     return this.evmDenomMap[chainName.toLowerCase()];
   }
 
-  async getDepositServiceContractAddress(chainName: string): Promise<string> {
-    if (!this.depositServiceContract[chainName]) {
-      this.depositServiceContract[chainName] = await this.getStaticInfo()
-        .then((body) => {
-          return body.assets.network[chainName.toLowerCase()]?.deposit_service;
-        })
-        .catch((e) => undefined);
-    }
-    return this.depositServiceContract[chainName];
-  }
-
   async getStaticInfo(): Promise<Record<string, any>> {
     if (!this.staticInfo) {
       this.staticInfo = await fetch(s3[this.environment])
         .then((res) => res.json())
-        .catch((e) => undefined);
+        .catch(() => undefined);
     }
     return this.staticInfo;
-  }
-
-  async validateChainIdentifiers(fromChain: string, toChain: string) {
-    const [fromChainValid, toChainValid] = await Promise.all([
-      validateChainIdentifier(fromChain, this.environment),
-      validateChainIdentifier(toChain, this.environment),
-    ]);
-    if (!fromChainValid.foundChain)
-      throw new Error(
-        `Invalid chain identifier for ${fromChain}. Did you mean ${fromChainValid.bestMatch}?`
-      );
-    if (!toChainValid.foundChain)
-      throw new Error(
-        `Invalid chain identifier for ${toChain}. Did you mean ${toChainValid.bestMatch}?`
-      );
-
-    return true;
   }
 }