+++ src/libs/TransactionRecoveryApi/AxelarGMPRecoveryAPI.ts
@@ -48,9 +48,8 @@ import {
   UnsupportedGasTokenError,
 } from "./constants/error";
 import { callExecute, CALL_EXECUTE_ERROR } from "./helpers";
-import { asyncRetry, sleep } from "../../utils";
+import { asyncRetry, sleep, throwIfInvalidChainIds } from "../../utils";
 import { BatchedCommandsResponse } from "@axelar-network/axelarjs-types/axelar/evm/v1beta1/query";
-import s3 from "./constants/s3";
 import { Interface } from "ethers/lib/utils";
 
 export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
@@ -180,11 +179,13 @@ export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
    */
   public async calculateNativeGasFee(
     txHash: string,
-    sourceChain: EvmChain,
-    destinationChain: EvmChain,
+    sourceChain: string,
+    destinationChain: string,
     gasTokenSymbol: GasToken | string,
     options: QueryGasFeeOptions
   ): Promise<string> {
+    await throwIfInvalidChainIds([sourceChain, destinationChain], this.environment);
+
     const provider = options.provider || getDefaultProvider(sourceChain, this.environment);
     const receipt = await provider.getTransactionReceipt(txHash);
     const paidGasFee = getNativeGasAmountFromTxReceipt(receipt) || "0";
@@ -208,6 +209,8 @@ export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
     gasTokenSymbol: GasToken | string,
     options: QueryGasFeeOptions
   ): Promise<string> {
+    await throwIfInvalidChainIds([sourceChain, destinationChain], this.environment);
+
     const provider = options.provider || getDefaultProvider(sourceChain, this.environment);
     const receipt = await provider.getTransactionReceipt(txHash);
     const paidGasFee = getGasAmountFromTxReceipt(receipt) || "0";
@@ -231,7 +234,10 @@ export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
     const evmWalletDetails = options?.evmWalletDetails || { useWindowEthereum: true };
     const signer = this.getSigner(chain, evmWalletDetails);
     const signerAddress = await signer.getAddress();
-    const gasReceiverAddress = await this.getGasReceiverContractAddress(chain);
+    const gasReceiverAddress = await this.axelarQueryApi.getContractAddressFromConfig(
+      chain,
+      "gas_service"
+    );
     const nativeGasTokenSymbol = NATIVE_GAS_TOKEN_SYMBOL[chain];
     const receipt = await signer.provider.getTransactionReceipt(txHash);
 
@@ -304,7 +310,10 @@ export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
     const evmWalletDetails = options?.evmWalletDetails || { useWindowEthereum: true };
     const signer = this.getSigner(chain, evmWalletDetails);
     const signerAddress = await signer.getAddress();
-    const gasReceiverAddress = await this.getGasReceiverContractAddress(chain);
+    const gasReceiverAddress = await this.axelarQueryApi.getContractAddressFromConfig(
+      chain,
+      "gas_service"
+    );
     const gasTokenContract = new ethers.Contract(gasTokenAddress, Erc20, signer.provider);
     const gasTokenSymbol = await gasTokenContract.symbol().catch(() => undefined);
 
@@ -444,8 +453,8 @@ export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
   }
 
   private async subtractGasFee(
-    sourceChain: EvmChain,
-    destinationChain: EvmChain,
+    sourceChain: string,
+    destinationChain: string,
     gasTokenSymbol: string,
     paidGasFee: string,
     options: QueryGasFeeOptions
@@ -462,7 +471,7 @@ export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
   }
 
   private getSigner(
-    chain: EvmChain,
+    chain: string,
     evmWalletDetails: EvmWalletDetails = { useWindowEthereum: true }
   ) {
     const { rpcMap, networkInfo } = rpcInfo[this.environment];
@@ -473,12 +482,5 @@ export class AxelarGMPRecoveryAPI extends AxelarRecoveryApi {
     };
     const evmClient = new EVMClient(evmClientConfig);
     return evmClient.getSigner();
-  }
-
-  public async getGasReceiverContractAddress(chainName: EvmChain): Promise<string> {
-    return await fetch(s3[this.environment])
-      .then((res) => res.json())
-      .then((body) => body.assets.network[chainName.toLowerCase()]?.gas_service)
-      .catch(() => "");
   }
 }