+++ test/integration/deposit-address/evm-evm.integration-spec.ts
@@ -1,8 +1,8 @@
-import { AxelarAssetTransfer, CHAINS } from "../../../src";
+import { AxelarAssetTransfer, CHAINS, Environment } from "../../../src";
 jest.setTimeout(20000);
 
 const sdk = new AxelarAssetTransfer({
-  environment: "testnet",
+  environment: Environment.TESTNET,
 });
 const evmAddress = "0xA57ADCE1d2fE72949E4308867D894CD7E7DE0ef2";
 const evmAsset = "uausdc";